+++ strato/CHANGELOG.md
@@ -25,7 +25,7 @@ so that they could be properly moved to their respective version's subsection.
 ### Fixed
 
 ### Removed
-- Removed 'block', 'blockGO' commands from blockapps-tools
+- Removed 'block', 'blockGO', 'canonRedis', 'compressRoundChanges' commands from blockapps-tools
 
 
 ## [11.1.0] - 3/28/2023